requirements.txt
numpy==1.24.3
sentence-transformers==2.2.2
faiss-cpu==1.7.4
openai==1.3.0
langchain==0.0.350
langchain-community==0.0.1
tiktoken==0.5.1
python-dotenv==1.0.0
pypdf2==3.0.1
beautifulsoup4==4.12.2
requests==2.31.0

